Make rmpass print an error message on false input.
The rmpass function asks the user whether he really wants to delete the
password store before actually performing the rm -rf. Although only Yes
or No should be accepted input, the function also interpreted wrong
input as a No. This is now changed and the function instead shows an
error message and aborts with a failure.